File Structure Refactor
