+++ src/components/AboutUs/AboutUs.jsx
@@ -1,6 +1,6 @@
 // Components
-import Header from "./Header/Header";
-import Main from "./Main/Main";
+import Header from "./Header";
+import Main from "./Main";
 
 export default function AboutUs() {
     return (
