+++ src/components/Global/Navbar.jsx
@@ -3,10 +3,10 @@ import { useState } from "react";
 import { Link } from "react-router-dom";
 
 // Assets
-import styles from "./navbar.module.css";
+import styles from "./styles/navbar.module.css";
 
 // Components
-import HamburgerMenu from "../HamburgerMenu/HamburgerMenu";
+import HamburgerMenu from "./HamburgerMenu";
 
 export default function Navbar() {
     const [hamburgerVisible, setHamburgerVisible] = useState(false);
