+++ src/components/Global/UnAuthenticated.jsx
@@ -3,10 +3,10 @@ import { useAuth0 } from "@auth0/auth0-react";
 import { User } from "react-feather";
 
 // Assets
-import styles from "./unAuthenticated.module.css";
+import styles from "./styles/unAuthenticated.module.css";
 
 // Components
-import Button from "../../Buttons/Button";
+import Button from "./Button";
 
 export default function UnAuthenticated() {
     const { loginWithRedirect } = useAuth0();
